chnaged place for the input system
Changed so that the input system is in gamemanger and also started on powerups
