+++ TeamWater/Models/Pickup_Thunder_Atlas.png.import
@@ -1,0 +1,36 @@
+[remap]
+
+importer="texture"
+type="CompressedTexture2D"
+uid="uid://bla1kr5hna105"
+path.s3tc="res://.godot/imported/Pickup_Thunder_Atlas.png-6939128ce72faf5dced7e6f21c3262c7.s3tc.ctex"
+metadata={
+"imported_formats": ["s3tc_bptc"],
+"vram_texture": true
+}
+generator_parameters={}
+
+[deps]
+
+source_file="res://Models/Pickup_Thunder_Atlas.png"
+dest_files=["res://.godot/imported/Pickup_Thunder_Atlas.png-6939128ce72faf5dced7e6f21c3262c7.s3tc.ctex"]
+
+[params]
+
+compress/mode=2
+compress/high_quality=false
+compress/lossy_quality=0.7
+compress/hdr_compression=1
+compress/normal_map=0
+compress/channel_pack=0
+mipmaps/generate=true
+mipmaps/limit=-1
+roughness/mode=0
+roughness/src_normal=""
+process/fix_alpha_border=true
+process/premult_alpha=false
+process/normal_map_invert_y=false
+process/hdr_as_srgb=false
+process/hdr_clamp_exposure=false
+process/size_limit=0
+detect_3d/compress_to=0
