+++ TeamWater/weapon_pickup.tscn
@@ -1,0 +1,20 @@
+[gd_scene load_steps=4 format=3 uid="uid://bc11ulb8i1rva"]
+
+[ext_resource type="Script" path="res://weapon_pickup.cs" id="1_kqn0j"]
+[ext_resource type="PackedScene" uid="uid://gf706oyjlei5" path="res://Models/Pickup_Sphere.glb" id="2_xtg1c"]
+
+[sub_resource type="BoxShape3D" id="BoxShape3D_7dv47"]
+
+[node name="Pickup" type="Area3D"]
+collision_layer = 4
+collision_mask = 9
+script = ExtResource("1_kqn0j")
+
+[node name="CollisionShape3D" type="CollisionShape3D" parent="."]
+shape = SubResource("BoxShape3D_7dv47")
+
+[node name="Pivot" type="Node3D" parent="."]
+
+[node name="Root Scene" parent="Pivot" instance=ExtResource("2_xtg1c")]
+
+[node name="AnimationPlayer" type="AnimationPlayer" parent="."]
